requirements.txt
asgiref==3.5.2
Django==4.1.2
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2==2.9.4
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
factory-boy==3.2.1
pytest~=7.1.3
pytest-django==4.5.2
pydantic==1.10.2